// angular/src/app/shared/enums/graph-colors.enum.ts
export enum GraphColor {
  FOREST_GREEN = '#228B22',
  LIGHT_GREEN = '#90EE90',
  DARK_GREY = '#A9A9A9',
  LIGHT_GREY = '#D3D3D3',
  INDIAN_RED = '#CD5C5C',
  LIGHT_INDIAN_RED = '#D77E7E',
  RED = '#E44D25',
  VISTA_BLUE = '#7AA3E5',
  MAROON = '#A8385D',
  CADET_BLUE = '#5F9EA0',
  LIGHT_SKY_BLUE = '#87CEFA',
  CHINESE_PURPLE = '#720789',
  RUBY_RED = '#A10A28',
  SCREAMIN_GREEN = '#7AFF59',
}